pyproject.toml
[tool.poetry]
name = "bitcoin-blockchain"
version = "0.1.0"
description = ""
authors = ["khaldi <jalalkhaldi3@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "==3.10.7"
pycryptodome = "==3.17"
jupyter = "^1.0.0"
python-socketio = {extras = ["client"], version = "^5.7.2"}
eventlet = "^0.33.3"
flask = "==2.1.2"
flask-classful = "^0.14.2"
werkzeug = "<=2.1.2"
flask-restx = ">=0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
